# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "投资收益统计系统 API",
    "description": "方案B — 资产快照 + 出入金 + 证券买卖 + 标的最新价 + 现金余额",
    "version": "1.0"
  },
  "paths": {
...
    "/api/portfolios": {
      "get": {
        "tags": [
          "portfolios"
        ],
        "summary": "List Portfolios",
        "operationId": "list_portfolios_api_portfolios_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PortfolioOut"
                  },
                  "type": "array",
                  "title": "Response List Portfolios Api Portfolios Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "portfolios"
        ],
        "summary": "Create Portfolio",
        "operationId": "create_portfolio_api_portfolios_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortfolioCreateReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/portfolios/{portfolio_id}": {
      "get": {
        "tags": [
          "portfolios"
        ],
        "summary": "Get Portfolio Detail",
        "operationId": "get_portfolio_detail_api_portfolios__portfolio_id__get",
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "parameters": [
          {
            "name": "portfolio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Portfolio Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "portfolios"
        ],
        "summary": "Patch Portfolio",
        "operationId": "patch_portfolio_api_portfolios__portfolio_id__patch",
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "parameters": [
          {
            "name": "portfolio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Portfolio Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortfolioPatchReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "portfolios"
        ],
        "summary": "Delete Portfolio",
        "operationId": "delete_portfolio_api_portfolios__portfolio_id__delete",
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "parameters": [
          {
            "name": "portfolio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Portfolio Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/portfolios/{portfolio_id}/data": {
      "delete": {
        "tags": [
          "portfolios"
        ],
        "summary": "Clear Data",
        "description": "清空组合所有数据（保留组合本身）。级联删除子表。",
        "operationId": "clear_data_api_portfolios__portfolio_id__data_delete",
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "parameters": [
          {
            "name": "portfolio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Portfolio Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClearDataOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/portfolios/{portfolio_id}/archive": {
      "patch": {
        "tags": [
          "portfolios"
        ],
        "summary": "Archive Portfolio",
        "description": "归档 / 取消归档（§4.2.2）。\n\narchived 缺省或 true → archivedAt = now；false → 置空。\n归档时若该组合为用户偏好默认组合，则同步置空（被隐藏的组合不能再当默认）。",
        "operationId": "archive_portfolio_api_portfolios__portfolio_id__archive_patch",
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "parameters": [
          {
            "name": "portfolio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Portfolio Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortfolioArchiveReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/users/preferences": {
      "get": {
        "tags": [
          "preference"
        ],
        "summary": "Get Preferences",
        "operationId": "get_preferences_api_users_preferences_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferenceOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "preference"
        ],
        "summary": "Patch Preferences",
        "operationId": "patch_preferences_api_users_preferences_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferenceOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ClearDataOut": {
        "properties": {
          "deletedCount": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Deletedcount"
          }
        },
        "type": "object",
        "required": [
          "deletedCount"
        ],
        "title": "ClearDataOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PortfolioArchiveReq": {
        "properties": {
          "archived": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived"
          }
        },
        "type": "object",
        "title": "PortfolioArchiveReq",
        "description": "归档请求：archived 缺省或 true → 归档；false → 取消归档。"
      },
      "PortfolioCreateReq": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "default": "CNY"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "PortfolioCreateReq"
      },
      "PortfolioOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "baseDate": {
            "type": "string",
            "format": "date",
            "title": "Basedate"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "archivedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archivedat"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "baseDate",
          "currency",
          "createdAt",
          "updatedAt"
        ],
        "title": "PortfolioOut"
      },
      "PortfolioPatchReq": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "title": "PortfolioPatchReq"
      },
...
      "PreferenceOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "defaultPortfolioId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Defaultportfolioid"
          },
          "defaultGranularity": {
            "type": "string",
            "title": "Defaultgranularity"
          },
          "defaultDateRange": {
            "type": "string",
            "title": "Defaultdaterange"
          },
          "aggregation": {
            "type": "string",
            "title": "Aggregation"
          },
          "weekStartsOn": {
            "type": "integer",
            "title": "Weekstartson"
          },
          "navDecimals": {
            "type": "integer",
            "title": "Navdecimals"
          },
          "xirrDecimals": {
            "type": "integer",
            "title": "Xirrdecimals"
          },
          "theme": {
            "type": "string",
            "title": "Theme"
          },
          "staleDays": {
            "type": "integer",
            "title": "Staledays"
          },
          "showLiquidated": {
            "type": "boolean",
            "title": "Showliquidated"
          },
          "costBasisView": {
            "type": "string",
            "title": "Costbasisview"
          },
          "cashHintOnCashflow": {
            "type": "boolean",
            "title": "Cashhintoncashflow"
          },
          "cashHintOnTrade": {
            "type": "boolean",
            "title": "Cashhintontrade"
          },
          "amountThousands": {
            "type": "boolean",
            "title": "Amountthousands"
          },
          "amountAbbrev": {
            "type": "boolean",
            "title": "Amountabbrev"
          },
          "dashboardLayout": {
            "type": "string",
            "title": "Dashboardlayout"
          }
        },
        "type": "object",
        "required": [
          "id",
          "defaultGranularity",
          "defaultDateRange",
          "aggregation",
          "weekStartsOn",
          "navDecimals",
          "xirrDecimals",
          "theme",
          "staleDays",
          "showLiquidated",
          "costBasisView",
          "cashHintOnCashflow",
          "cashHintOnTrade",
          "amountThousands",
          "amountAbbrev",
          "dashboardLayout"
        ],
        "title": "PreferenceOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "JWT-auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}